Move location of utilities outside workflows' folder.
The utils folder will host utilities that are not only for the
workflows.
