+++ aiida_siesta/workflows/eos.py
@@ -4,7 +4,7 @@ from aiida.orm import Float
 from aiida_siesta.calculations.tkdict import FDFDict
 from aiida_siesta.workflows.base import SiestaBaseWorkChain
 
-from .utils.iterate_absclass import BaseIterator
+from ..utils.iterate_absclass import BaseIterator
 
 
 @calcfunction
@@ -308,5 +308,5 @@ class EqOfStateFixedCellShape(BaseIterator):
 
     @classmethod
     def inputs_generator(cls):  # pylint: disable=no-self-argument,no-self-use
-        from aiida_siesta.workflows.utils.inputs_generators import EosWorkChainInputsGenerator
+        from aiida_siesta.utils.inputs_generators import EosWorkChainInputsGenerator
         return EosWorkChainInputsGenerator(cls)
