+++ aiida_siesta/workflows/stm.py
@@ -310,5 +310,5 @@ class SiestaSTMWorkChain(WorkChain):
 
     @classmethod
     def inputs_generator(cls):  # pylint: disable=no-self-argument,no-self-use
-        from aiida_siesta.workflows.utils.inputs_generators import StmWorkChainInputsGenerator
+        from aiida_siesta.utils.inputs_generators import StmWorkChainInputsGenerator
         return StmWorkChainInputsGenerator(cls)
