+++ tests/utils/fixtures/protocols/registries/custom_prot.yaml
